+++ Tests/Unit/Domain/Model/Event/EventTopicTest.php
@@ -528,4 +528,24 @@ final class EventTopicTest extends UnitTestCase
 
         $this->subject->getPriceByPriceCode(Price::PRICE_EARLY_BIRD);
     }
+
+    /**
+     * @test
+     */
+    public function getRawDataInitiallyReturnsNull(): void
+    {
+        self::assertNull($this->subject->getRawData());
+    }
+
+    /**
+     * @test
+     */
+    public function setRawDataSetsRawData(): void
+    {
+        $rawData = ['uid' => 5, 'title' => 'foo'];
+
+        $this->subject->setRawData($rawData);
+
+        self::assertSame($rawData, $this->subject->getRawData());
+    }
 }