+++ Tests/Unit/Domain/Model/Event/SingleEventTest.php
@@ -1060,4 +1060,24 @@ final class SingleEventTest extends UnitTestCase
 
         self::assertFalse($this->subject->allowsUnlimitedRegistrations());
     }
+
+    /**
+     * @test
+     */
+    public function getRawDataInitiallyReturnsNull(): void
+    {
+        self::assertNull($this->subject->getRawData());
+    }
+
+    /**
+     * @test
+     */
+    public function setRawDataSetsRawData(): void
+    {
+        $rawData = ['uid' => 5, 'title' => 'foo'];
+
+        $this->subject->setRawData($rawData);
+
+        self::assertSame($rawData, $this->subject->getRawData());
+    }
 }